+++ zarrs_metadata_ext/src/data_type.rs
@@ -2,9 +2,16 @@
 //!
 //! See <https://zarr-specs.readthedocs.io/en/latest/v3/core/index.html#array-metadata-data-type>.
 
-pub mod numpy_datetime64;
-pub mod numpy_timedelta64;
 pub mod optional;
 
-mod numpy_time_unit;
-pub use numpy_time_unit::NumpyTimeUnit;
+/// Re-export `numpy.datetime64` data type metadata from `zarrs_metadata_numpy_time`.
+pub mod numpy_datetime64 {
+    pub use zarrs_metadata_numpy_time::NumpyDateTime64DataTypeConfigurationV1;
+}
+
+/// Re-export `numpy.timedelta64` data type metadata from `zarrs_metadata_numpy_time`.
+pub mod numpy_timedelta64 {
+    pub use zarrs_metadata_numpy_time::NumpyTimeDelta64DataTypeConfigurationV1;
+}
+
+pub use zarrs_metadata_numpy_time::NumpyTimeUnit;