+++ zarrs_metadata_ext/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `NumpyTimeUnit`, `NumpyDateTime64DataTypeConfigurationV1`, and `NumpyTimeDelta64DataTypeConfigurationV1` are now re-exported from `zarrs_metadata_numpy_time` 0.1.0
+
 ## [0.4.0] - 2025-12-31
 
 ### Removed